Addressng chsh fail due to /etc/passwd permissions
Since maas containers are running as non-privileged,
the maas root user might not be able to get lock on
the host's /etc/paasword to change shell for the maas
user. Added retry/sleep to eventually get the lock and
make the shell update for the user.
Also, using usermod instead of chsh.

Change-Id: Iaba6276bd6d54b22980aef8c6f7c5c113d0f2fff