MAAS chart: reduce syslog startup spam
When the MAAS syslog pod starts, it polls continuously until the log
file exists, generating a message every 10 seconds. However, rsyslogd
won't create the file until it receives the first message, which could
take a while.

This change will create an empty file if none exists prior to starting
the rsyslogd service.

Previous comments indicate some concerns about a race condition, and it
is possible that there are some circumstances when the file may go away
and come back, so the polling loop is left in place.

Change-Id: Ic56faf718038c5d17ab9353399a94ec74e91f8d0