maas-syslog fix: eliminate duplicate messages
A recent change[0] to allow customization of the log level inadvertently
resulted in most messages being logged twice - once if they matched the
severity constraint, and again for all non-local messages, which for the
intended use case is all of them.

This change corrects the rsyslog.conf to drop local messages, and log
the remainder at the configured severity level. It also removes the
"$RepeatedMsgReduction on" parameter, which may have partially masked
the issue, and whose use is not advised.[1]

Change-Id: Ib15f82d9e1c7cef7d6085d6a215354b064aa09bb
0: https://opendev.org/airship/maas/commit/e22afb6e953bd8ab8acaa6ebbc04d72c50cd6fb2
1: https://www.rsyslog.com/doc/v8-stable/configuration/action/rsconf1_repeatedmsgreduction.html