Disable creation of swap file
By default, curtin creates a swap file of up to 8GB. When swap is later
disabled, there is still a /swap.img file left hanging around that needs
to be cleaned up.

This change sets the size to 0 to disable the creation of the swap file
in the first place.

https://curtin.readthedocs.io/en/latest/topics/config.html#swap

Change-Id: I9e1e5f67007ae3c49617525e989b27e123b69d53