Allow additional late_commands in curtin userdata
This change allows extra late_commands to be added to the curtin
userdata, which are executed before the node is rebooted at the end of
the deployment. This can be useful to install packages or perform other
customization.

One sample use-case is the installation of specific kernel module
packages that match the target kernel image, in cases where the
ephemeral environment uses a different kernel version.

Change-Id: I80084c544f6a7dafd6aa84c8041cf86bdc3b9f4b