Improved MAAS helm test reliability and logging
This PS adds more robust checks to the helm test
script to ensure that the MAAS services are
properly deployed and functional before marking
the test as successful. It includes additional
logging for easier debugging in case of failures.

In addition it updated import-boot-resources.sh
to add stall detection and restart logic to
handle cases where the boot image import process
may hang indefinitely.

Change-Id: Ic1ccf8ec60bb0b25988bc82c589537933f9f74c1
Signed-off-by: Sergiy Markin <smarkin@mirantis.com>